+++ CHANGELOG.md
@@ -48,6 +48,8 @@ This format is based on [Keep A Changelog](https://keepachangelog.com/en/1.0.0).
     - `initLovelaceAssertValue`
     - `initLovelaceAssertValueWith`
     - `withCollateral`
+  - Initialising wallets with staking keys.
+  - Access to initialised wallets via `WalletLookups`
 
 ## 0.1 -- 2022-02-14